+++ torch_fidelity/deprecations.py
@@ -1,0 +1,17 @@
+DEPRECATIONS = {
+    "kid_degree": {
+        "new_name": "kid_kernel_poly_degree",
+        "since": "0.4.0",
+        "reason": "Supporting various kernel functions, including RBF, to enable computing the metric proposed in https://arxiv.org/pdf/2401.09603.pdf",
+    },
+    "kid_gamma": {
+        "new_name": "kid_kernel_poly_gamma",
+        "since": "0.4.0",
+        "reason": "Supporting various kernel functions, including RBF, to enable computing the metric proposed in https://arxiv.org/pdf/2401.09603.pdf",
+    },
+    "kid_coef0": {
+        "new_name": "kid_kernel_poly_coef0",
+        "since": "0.4.0",
+        "reason": "Supporting various kernel functions, including RBF, to enable computing the metric proposed in https://arxiv.org/pdf/2401.09603.pdf",
+    },
+}